IPUMS

Documentation for IPUMS.

IPUMS.IPUMSSource

IPUMS.ipums_data_collections

IPUMS.IPUMSSource — Type

IPUMSSource(
    proj_name::String, 
    url_name::String, 
    collection_type::String, 
    code_for_api::String = "", 
    api_support::Bool = false, 
    home_url::String = ""
)

A struct representing sources that IPUMS provides.

Arguments

proj_name::String - Name of the IPUMS project.

url_name::String - Name of the project as used in that project's website URL.

collection_type::String - Either "microdata" or "aggregate data" indicating the type of data this collection provides.

Keyword Arguments

code_for_api::String - The name of the project used when interacting with the IPUMS API (for collections that are supported by the API). (Default: "")

api_support::Bool - Logical indicating whether the collection is supported by the IPUMS API. (Default: false)

home_url::String - URL for the project's homepage. (Default: "")

Returns

IPUMSSource object specifying the previous parameters

Example

julia> IPUMSSource(
    proj_name = "IPUMS USA",
    url_name = "usa",
    collection_type = "microdata",
    api_support = true
)

IPUMS.IPUMSSource("IPUMS USA", "usa", "microdata", "", true, "")

IPUMS.ipums_data_collections — Method

ipums_data_collections()

List IPUMS data collections with their corresponding codes used by the IPUMS API. Unlisted data collections are not yet supported by the IPUMS API.

Returns

DataFrame with four columns containing the full collection name, the type of data the collection provides, the collection code used by the IPUMS API, and the status of API support for the collection.

Example

julia> ipums_data_collections()

 Row │ collection_name      collection_type  code_for_api  api_support 
     │ String               String           String        Bool        
─────┼─────────────────────────────────────────────────────────────────
   1 │ IPUMS USA            microdata                             true
   2 │ IPUMS CPS            microdata                             true
   3 │ IPUMS International  microdata        ipumsi               true
   ...